Modified src/main.c

+++ src/main.c
@@ -22,9 +22,9 @@
 #pragma config BORV = LO        // Brown-out Reset Voltage Selection (Brown-out Reset Voltage (Vbor), low trip point selected.)
 #pragma config LVP = OFF        // Low-Voltage Programming Enable (High-voltage on MCLR/VPP must be used for programming)
 
-void delay_10ms(unsigned int num)
+void delay_10ms(uint16_t num)
 {
-     uint8_t lc;
+     uint16_t lc;
 
      for (lc = 0 ; lc < num ; lc++)
      {
@@ -35,17 +35,17 @@ void delay_10ms(unsigned int num)
 void main()
 {
      OSCCON = 0b01101010;      // INTOSC: 4MHz
-     ANSELA = 0b00000000;      // Analog: All disable
-     TRISA  = 0b00001000;      // I/O: All out
+     ANSELA = 0b00000000;      // Analog: All Disable
+     TRISA  = 0b00001000;      // I/O: All OUT
      PORTA  = 0b00000000;      // Port: All LOW
 
      // Every second LED Blink
      while(1)
      {
           RA0 = 1;             // Pin No.7(RA0): HIGH
-          delay_10ms(100);     // Wait 1 sec
+          delay_10ms(100);     // Wait 1.0 sec
           RA0 = 0;             // Pin No.7(RA0): LOW
-          delay_10ms(100);     // Wait 1 sec
+          delay_10ms(100);     // Wait 1.0 sec
     }
 }
